+++ .dev/update-version/cargo.mjs
@@ -1,6 +1,6 @@
 
 /**
- * Update-version.mjs.
+ * Update cargo version.
  *
  * @description Update cargo version.
  */ 
@@ -8,35 +8,30 @@
 import {
 	readFile, writeFile, 
 } from 'fs/promises'
-import path from 'path'
+import {
+	joinPath, pkg, 
+} from '../core/main.mjs'
 
-const updateCargoVersion = async () => {
+const workspacePath = pkg.dir
+const pkgAppPath    = joinPath( workspacePath, 'packages', 'app' )
+
+export const updateCargoVersion = async () => {
 
 	try {
 
-		// Leer package.json
-		const packageJsonPath = path.join( process.cwd(), 'package.json' )
+		const packageJsonPath = joinPath( pkgAppPath, 'package.json' )
 		const packageJson     = JSON.parse( await readFile( packageJsonPath, 'utf-8' ) )
 		const newVersion      = packageJson.version
 
-		// Leer Cargo.toml
-		const cargoTomlPath  = path.join( process.cwd(), 'src-tauri', 'Cargo.toml' )
+		const cargoTomlPath  = joinPath( pkgAppPath, 'src-tauri', 'Cargo.toml' )
 		let cargoTomlContent = await readFile( cargoTomlPath, 'utf-8' )
 
-		// Buscar y reemplazar la línea con la versión
 		const versionRegex = /^version\s*=\s*"(.*)"/m
-		if ( versionRegex.test( cargoTomlContent ) ) {
-
+		if ( versionRegex.test( cargoTomlContent ) ) 
 			cargoTomlContent = cargoTomlContent.replace( versionRegex, `version = "${newVersion}"` )
-		
-		} else {
-
-			// Si no existe el campo, lo agregamos al final del archivo
+		else 
 			cargoTomlContent += `\n[package]\nversion = "${newVersion}"\n`
-		
-		}
 
-		// Escribir el archivo actualizado
 		await writeFile( cargoTomlPath, cargoTomlContent, 'utf-8' )
 
 		console.log( `Version updated to ${newVersion} in Cargo.toml` )
@@ -48,5 +43,3 @@ const updateCargoVersion = async () => {
 	}
 
 }
-
-updateCargoVersion()